// src/api/generated/model/baseRoomDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

export interface BaseRoomDto {
  /** Unique identifier */
  id: string;
  /** Timestamp of when the record was created */
  createdAt: string;
  /** Timestamp of the last update of the record */
  updatedAt: string;
  /** Название комнаты */
  name: string;
  /** ID группы подписчиков в Сенлере (ожидают) */
  pendingSubscriptionId: number;
  /** ID группы подписчиков в Сенлере (одобренные) */
  approvedSubscriptionId: number;
  /** ID группы подписчиков в Сенлере (отклоненные) */
  rejectedSubscriptionId: number;
  /**
   * ID бота в Senler для уведомления об успешно принятом креативе
   * @nullable
   */
  notificationCreativeTaskApprovedBotId: string | null;
  /**
   * ID бота в Senler для уведомления об отклоненном креативе
   * @nullable
   */
  notificationCreativeTaskRejectedBotId: string | null;
  /**
   * URL вебхука
   * @nullable
   */
  webhookUrl: string | null;
  /** Секретный ключ */
  secretKey: string;
  /** Флаг удаления */
  isDeleted: boolean;
  /** Id сенлер проекта */
  projectId: string;
}