// src/api/generated/model/updateRoomOrdProfileRequestDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

export interface UpdateRoomOrdProfileRequestDto {
  /**
   * ФИО или наименование организации
   * @maxLength 255
   */
  name?: string;
  /** Номер телефона с кодом страны, например +79876543210 */
  phone?: string;
  /** Российский ИНН; обязателен для physical, ip, juridical */
  inn?: string;
  /** Адрес регистрации; опционален для совместимости со старыми запросами */
  address?: string;
  /** Цифровой код страны ISO 3166; обязателен для иностранных контрагентов */
  foreignOksmCountryCode?: string;
  /** Номер карты или счёта; обязателен для иностранного физлица */
  foreignEpaymentMethod?: string;
  /** Регистрационный номер; обязателен для иностранного юрлица */
  foreignRegistrationNumber?: string;
  /** Иностранный ИНН: обязателен для иностранного юрлица; для иностранного физлица принимается без сохранения */
  foreignInn?: string;
}